refactor: lift the target aiming out of the palette
Aiming at a place in the diagram — holding a target, stepping it with the
arrows, previewing it on the canvas, naming it for assistive tech — was
written inside the stencil palette, so only the palette could use it. The
canvas commands each settled for a cruder rule instead: `shift+cmd+enter`
inserts a lane after the selected one, and `cmd+enter` appends to the end,
neither of them aimed or previewed.

It is `useTargetAim()` now, with a command-shaped surface — `begin`,
`end`, `step`, `target`, `name` — leaving the palette to wire it to the
buttons. Nothing changes for the user; the aiming is simply reachable
from somewhere other than the stencil.

The move also brought one rule back into the open: a group has nowhere to
aim, since it is never embedded. That was buried in the preview effect
and is now `aimsAtSomething()`, asked where the question belongs.
